refactor: put everything into different folders
