+++ frontend/app/mocks/auth/refresh/route.ts
@@ -1,4 +1,4 @@
-import { access_token, refresh_token } from 'features/auth'
+import { access_token, refresh_token } from 'features/auth/mocks'
 import { NextResponse } from 'next/server'
 
 export async function POST(req: Request) {
