+++ frontend/features/auth/index.ts
@@ -1,2 +1,1 @@
 export * from './utils'
-export * from './auth'
