+++ frontend/features/auth/types/authPayload.ts
@@ -1,0 +1,4 @@
+export interface AuthPayload {
+  username: string
+  password: string
+}
